Problem with Numpy
Numpy would not initialise on my machine - there was a conflict and I couldn't import it. So I couldn't verify if this code works or not - I used the resource at https://www.geeksforgeeks.org/how-to-plot-normal-distribution-over-histogram-in-python/ to get to my starting point, but then couldn't go any further. Original error code was "import "numpy" could not be resolved pylace(reportmissingimports)" so after a bit of research i saw that I could manually add it, this didn't work either. Numpy it seems is installed on my machine in the Anaconda folder, but I couldn't get it to pull over to VSCode.